MISC: Refactor Person.ts and Sleeve.ts
